[Upd] Refactor
Validation of questions were separated from validator factory.
